% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_strings.R
\name{extract_r2}
\alias{extract_r2}
\title{Wrapper to extract R-squared from a linear regression model}
\usage{
extract_r2(mod)
}
\arguments{
\item{mod}{A linear regression model}
}
\value{
Model R-squared
}
\description{
This function takes an object produced by stats::lm() and returns
the model R-squared
}
