% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_strings.R
\name{extract_stats}
\alias{extract_stats}
\title{Wrapper to format raw ANOVA results for APA reporting}
\usage{
extract_stats(aov_summary_obj, predictor = NA)
}
\arguments{
\item{aov_summary_obj}{The ANOVA summary object}

\item{predictor}{The row or variable name to format output for}
}
\value{
An APA-formatted string of ANOVA results
}
\description{
This function takes the summary object from an ANOVA object and feeds it
to the format_anova_string function
}
