% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_strings.R
\name{format_anova_string}
\alias{format_anova_string}
\title{Extract statistics and format raw ANOVA results for APA reporting in markdown}
\usage{
format_anova_string(aov_summary_obj, predictor = NA)
}
\arguments{
\item{aov_summary_obj}{The ANOVA summary object}

\item{predictor}{The row or variable name to format output for (defaults to last row)}
}
\value{
An APA-formatted string of ANOVA results
}
\description{
This function takes the summary object from an ANOVA object and feeds it
to the format_anova_string function.
Accepted ANOVA objects include:
stats::aov()
summary(stats::aov())
rstatix::get_anova_table()
apaTables::apa.aov.table()
}
