% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_strings.R
\name{format_corr_apa}
\alias{format_corr_apa}
\title{Format results of a bivariate correlation}
\usage{
format_corr_apa(r, dof, p, as.markdown = T)
}
\arguments{
\item{r}{value of correlation to be formatted}

\item{dof}{degrees of freedom for the correlation}

\item{p}{p-value of correlation}

\item{as.markdown}{optional, indicates whether the resulting string should be formatted for markdown. Default is TRUE}
}
\value{
A formatted bivariate correlation statistic
}
\description{
This function takes the results of a bivariate correlation and formats it for
easy markdown outputs
}
