% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_strings.R
\name{format_pairwise_comparison}
\alias{format_pairwise_comparison}
\title{Format results of a pairwise comparison}
\usage{
format_pairwise_comparison(
  t_stat,
  df,
  p_val,
  mdiff = "none",
  lci = "none",
  uci = "none",
  cohen_d = "none"
)
}
\arguments{
\item{t_stat}{t-statistic}

\item{df}{degrees of freedom}

\item{p_val}{p-value of t-statistic}

\item{mdiff}{optional, the mean difference of the comparison}

\item{lci}{optional, the lower end of the confidence interval on the comparison}

\item{uci}{optional, the upper end of the confidence interval on the comparison}

\item{cohen_d}{optional, Cohen's d standardized effect size}
}
\value{
A formatted t-statistic
}
\description{
This function takes the results of a pairwise comparison and formats it for
easy markdown outputs
}
