% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_strings.R
\name{format_pval_apa}
\alias{format_pval_apa}
\title{Format a p-value}
\usage{
format_pval_apa(p_val)
}
\arguments{
\item{p_val}{P-value to be formatted}
}
\value{
A formatted p-value string
}
\description{
This function takes a p-value (numeric or character) and formats it so that
values less than 0.001 are treated as <0.001
}
