% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_strings.R
\name{format_sig_stars}
\alias{format_sig_stars}
\title{Get p-value significance stars}
\usage{
format_sig_stars(p_val)
}
\arguments{
\item{p_val}{P-value to get stars for}
}
\value{
A string of significance asterisks
}
\description{
This function takes a p-value and returns a string containing significance asterisks
Returns nothing if p > 0.05
}
