% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_strings.R
\name{format_tstat_apa}
\alias{format_tstat_apa}
\title{Format results of a t-test}
\usage{
format_tstat_apa(t_result)
}
\arguments{
\item{t_result}{t-test result to be formatted}
}
\value{
A formatted t-statistic
}
\description{
This function takes the results of a t-test and formats it for
easy markdown outputs
}
