% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_strings.R
\name{format_anova_string}
\alias{format_anova_string}
\title{Extract statistics and format raw ANOVA results for APA reporting in markdown}
\usage{
format_anova_string(aov_summary_obj, as.markdown = T, get.all = T)
}
\arguments{
\item{aov_summary_obj}{The ANOVA summary object}

\item{as.markdown}{optional, indicates whether the resulting string should be formatted for markdown. Default is TRUE}

\item{get.all}{deprecated, no longer an option but retaining in the function prevents errors from prior versions}
}
\value{
An APA-formatted string (or strings) of ANOVA results
}
\description{
This function takes the summary object from an ANOVA object and feeds it
to the format_anova_string function.
Accepted ANOVA objects include:
stats::aov()
summary(stats::aov())
rstatix::get_anova_table()
apaTables::apa.aov.table()
}
