% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_strings.R
\name{format_r2}
\alias{format_r2}
\title{Wrapper to extract R-squared from a linear regression model}
\usage{
format_r2(mod, as.markdown = T)
}
\arguments{
\item{mod}{A linear regression model or the R-squared from a regression model}

\item{as.markdown}{optional, indicates whether the resulting string should be formatted for markdown. Default is TRUE}
}
\value{
Formatted string containing model R-squared
}
\description{
This function takes an object produced by stats::lm() and returns
the model R-squared as a string formaatted for markdown output
}
