% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_strings.R
\name{pval_fmt}
\alias{pval_fmt}
\title{Helper function to format p-value}
\usage{
pval_fmt(p_val, markdown = F)
}
\arguments{
\item{p_val}{P-value to be formatted}
}
\value{
A formatted p-value string
}
\description{
This is the base function that format_pval_apa calls.
}
